[IMP] Created tests_rest_tasks.py
