+++ tests/tests_rest.py
@@ -532,32 +532,3 @@ class TestsRest(TestCase):
         response = user.delete(f'/api/v2/assets/{asset_identifier}')
         self.assertEqual(403, response.status_code)
 
-    def test_get_task_should_return_403_when_user_has_insufficient_rights(self):
-        case_identifier = self._subject.create_dummy_case()
-        user = self._subject.create_dummy_user()
-        body = {'task_assignees_id': [1], 'task_description': '', 'task_status_id': 1, 'task_tags': '', 'task_title': 'dummy title', 'custom_attributes': {}}
-        response = self._subject.create(f'/api/v2/cases/{case_identifier}/tasks',  body).json()
-        task_identifier = response['id']
-        body = {
-            'cases_list': [_INITIAL_DEMO_CASE_IDENTIFIER],
-            'access_level': _CASE_ACCESS_LEVEL_FULL_ACCESS
-        }
-        self._subject.create(f'/manage/users/{user.get_identifier()}/cases-access/update', body)
-
-        response = user.get(f'/api/v2/tasks/{task_identifier}')
-        self.assertEqual(403, response.status_code)
-
-    def test_delete_task_should_return_403_when_user_has_insufficient_rights(self):
-        case_identifier = self._subject.create_dummy_case()
-        user = self._subject.create_dummy_user()
-        body = {'task_assignees_id': [1], 'task_description': '', 'task_status_id': 1, 'task_tags': '', 'task_title': 'dummy title', 'custom_attributes': {}}
-        response = self._subject.create(f'/api/v2/cases/{case_identifier}/tasks',  body).json()
-        task_identifier = response['id']
-        body = {
-            'cases_list': [_INITIAL_DEMO_CASE_IDENTIFIER],
-            'access_level': _CASE_ACCESS_LEVEL_FULL_ACCESS
-        }
-        self._subject.create(f'/manage/users/{user.get_identifier()}/cases-access/update', body)
-
-        response = user.delete(f'/api/v2/tasks/{task_identifier}')
-        self.assertEqual(403, response.status_code)
